# kk/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for SafeDrive cloud storage
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type FileItemFileType =
  (typeof FileItemFileType)[keyof typeof FileItemFileType];

export const FileItemFileType = {
  image: "image",
  video: "video",
  document: "document",
  other: "other",
} as const;

export interface FileItem {
  id: number;
  userId: string;
  name: string;
  size: number;
  contentType: string;
  fileType: FileItemFileType;
  objectPath: string;
  createdAt: string;
}

export type CreateFileBodyFileType =
  (typeof CreateFileBodyFileType)[keyof typeof CreateFileBodyFileType];

export const CreateFileBodyFileType = {
  image: "image",
  video: "video",
  document: "document",
  other: "other",
} as const;

export interface CreateFileBody {
  name: string;
  size: number;
  contentType: string;
  fileType: CreateFileBodyFileType;
  objectPath: string;
}

export interface FileStats {
  totalFiles: number;
  totalSize: number;
  imageCount: number;
  videoCount: number;
  documentCount: number;
  otherCount: number;
}

export interface RequestUploadUrlBody {
  name: string;
  size: number;
  contentType: string;
}

export interface RequestUploadUrlResponse {
  uploadURL: string;
  objectPath: string;
}

export type ListFilesParams = {
  type?: ListFilesType;
  search?: string;
};

export type ListFilesType = (typeof ListFilesType)[keyof typeof ListFilesType];

export const ListFilesType = {
  image: "image",
  video: "video",
  document: "document",
  other: "other",
} as const;