# kk/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for SafeDrive cloud storage
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: files
    description: File management
  - name: storage
    description: Object storage upload/serve
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /files:
    get:
      operationId: listFiles
      tags: [files]
      summary: List all files for the current user
      parameters:
        - in: query
          name: type
          schema:
            type: string
            enum: [image, video, document, other]
          required: false
        - in: query
          name: search
          schema:
            type: string
          required: false
      responses:
        "200":
          description: List of files
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FileItem"
        "401":
          description: Unauthorized
    post:
      operationId: createFile
      tags: [files]
      summary: Create a file record after upload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFileBody"
      responses:
        "201":
          description: Created file record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileItem"
        "401":
          description: Unauthorized

  /files/{id}:
    get:
      operationId: getFile
      tags: [files]
      summary: Get a single file by ID
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      responses:
        "200":
          description: File record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileItem"
        "404":
          description: Not found
    delete:
      operationId: deleteFile
      tags: [files]
      summary: Delete a file record
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
        "404":
          description: Not found

  /files/stats:
    get:
      operationId: getFileStats
      tags: [files]
      summary: Get storage usage stats for the current user
      responses:
        "200":
          description: Storage stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileStats"
        "401":
          description: Unauthorized

  /files/recent:
    get:
      operationId: getRecentFiles
      tags: [files]
      summary: Get recently uploaded files
      responses:
        "200":
          description: Recent files
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FileItem"
        "401":
          description: Unauthorized

  /storage/uploads/request-url:
    post:
      operationId: requestUploadUrl
      tags: [storage]
      summary: Request a presigned upload URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestUploadUrlBody"
      responses:
        "200":
          description: Presigned upload URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestUploadUrlResponse"
        "401":
          description: Unauthorized

  /storage/objects/{objectPath}:
    get:
      operationId: getObject
      tags: [storage]
      summary: Serve a stored object
      parameters:
        - in: path
          name: objectPath
          schema:
            type: string
          required: true
      responses:
        "200":
          description: Object content

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    FileItem:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        name:
          type: string
        size:
          type: integer
        contentType:
          type: string
        fileType:
          type: string
          enum: [image, video, document, other]
        objectPath:
          type: string
        createdAt:
          type: string
      required:
        - id
        - userId
        - name
        - size
        - contentType
        - fileType
        - objectPath
        - createdAt

    CreateFileBody:
      type: object
      properties:
        name:
          type: string
        size:
          type: integer
        contentType:
          type: string
        fileType:
          type: string
          enum: [image, video, document, other]
        objectPath:
          type: string
      required:
        - name
        - size
        - contentType
        - fileType
        - objectPath

    FileStats:
      type: object
      properties:
        totalFiles:
          type: integer
        totalSize:
          type: integer
        imageCount:
          type: integer
        videoCount:
          type: integer
        documentCount:
          type: integer
        otherCount:
          type: integer
      required:
        - totalFiles
        - totalSize
        - imageCount
        - videoCount
        - documentCount
        - otherCount

    RequestUploadUrlBody:
      type: object
      properties:
        name:
          type: string
        size:
          type: integer
        contentType:
          type: string
      required:
        - name
        - size
        - contentType

    RequestUploadUrlResponse:
      type: object
      properties:
        uploadURL:
          type: string
        objectPath:
          type: string
      required:
        - uploadURL
        - objectPath