# kk/lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for SafeDrive cloud storage
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all files for the current user
 */
export const ListFilesQueryParams = zod.object({
  type: zod.enum(["image", "video", "document", "other"]).optional(),
  search: zod.coerce.string().optional(),
});

export const ListFilesResponseItem = zod.object({
  id: zod.number(),
  userId: zod.string(),
  name: zod.string(),
  size: zod.number(),
  contentType: zod.string(),
  fileType: zod.enum(["image", "video", "document", "other"]),
  objectPath: zod.string(),
  createdAt: zod.string(),
});
export const ListFilesResponse = zod.array(ListFilesResponseItem);

/**
 * @summary Create a file record after upload
 */
export const CreateFileBody = zod.object({
  name: zod.string(),
  size: zod.number(),
  contentType: zod.string(),
  fileType: zod.enum(["image", "video", "document", "other"]),
  objectPath: zod.string(),
});

/**
 * @summary Get a single file by ID
 */
export const GetFileParams = zod.object({
  id: zod.coerce.number(),
});

export const GetFileResponse = zod.object({
  id: zod.number(),
  userId: zod.string(),
  name: zod.string(),
  size: zod.number(),
  contentType: zod.string(),
  fileType: zod.enum(["image", "video", "document", "other"]),
  objectPath: zod.string(),
  createdAt: zod.string(),
});

/**
 * @summary Delete a file record
 */
export const DeleteFileParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Get storage usage stats for the current user
 */
export const GetFileStatsResponse = zod.object({
  totalFiles: zod.number(),
  totalSize: zod.number(),
  imageCount: zod.number(),
  videoCount: zod.number(),
  documentCount: zod.number(),
  otherCount: zod.number(),
});

/**
 * @summary Get recently uploaded files
 */
export const GetRecentFilesResponseItem = zod.object({
  id: zod.number(),
  userId: zod.string(),
  name: zod.string(),
  size: zod.number(),
  contentType: zod.string(),
  fileType: zod.enum(["image", "video", "document", "other"]),
  objectPath: zod.string(),
  createdAt: zod.string(),
});
export const GetRecentFilesResponse = zod.array(GetRecentFilesResponseItem);

/**
 * @summary Request a presigned upload URL
 */
export const RequestUploadUrlBody = zod.object({
  name: zod.string(),
  size: zod.number(),
  contentType: zod.string(),
});

export const RequestUploadUrlResponse = zod.object({
  uploadURL: zod.string(),
  objectPath: zod.string(),
});

/**
 * @summary Serve a stored object
 */
export const GetObjectParams = zod.object({
  objectPath: zod.coerce.string(),
});