# kk/lib/api-zod/src/generated/types/createFileBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for SafeDrive cloud storage
 * OpenAPI spec version: 0.1.0
 */
import type { CreateFileBodyFileType } from "./createFileBodyFileType";

export interface CreateFileBody {
  name: string;
  size: number;
  contentType: string;
  fileType: CreateFileBodyFileType;
  objectPath: string;
}