# kk/lib/api-zod/src/generated/types/fileItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for SafeDrive cloud storage
 * OpenAPI spec version: 0.1.0
 */
import type { FileItemFileType } from "./fileItemFileType";

export interface FileItem {
  id: number;
  userId: string;
  name: string;
  size: number;
  contentType: string;
  fileType: FileItemFileType;
  objectPath: string;
  createdAt: string;
}