# kk/lib/api-zod/src/generated/types/requestUploadUrlBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for SafeDrive cloud storage
 * OpenAPI spec version: 0.1.0
 */

export interface RequestUploadUrlBody {
  name: string;
  size: number;
  contentType: string;
}